# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Forum Furiosum API
  version: 1.0.0
  description: REST API for Forum Furiosum web forum
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ff_session
...
paths:
...
  /threads:
    get:
      summary: List threads
      tags: [Threads]
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated thread list
    post:
      summary: Create thread
      tags: [Threads]
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, title, body]
              properties:
                categoryId:
                  type: string
                title:
                  type: string
                  minLength: 3
                  maxLength: 120
                body:
                  type: string
                  minLength: 1
                  maxLength: 10000
      responses:
        '201':
          description: Thread created
        '401':
          description: Unauthorized
  /threads/{id}:
    get:
      summary: Get thread
      tags: [Threads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thread object
        '404':
          description: Not found
    patch:
      summary: Edit thread
      tags: [Threads]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 120
                body:
                  type: string
                  minLength: 1
                  maxLength: 10000
                reason:
                  type: string
      responses:
        '200':
          description: Updated thread
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      summary: Delete thread
      tags: [Threads]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
  /threads/{id}/posts:
    get:
      summary: List posts in thread
      tags: [Posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated post list
        '404':
          description: Thread not found
    post:
      summary: Create post in thread
      tags: [Posts]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 10000
      responses:
        '201':
          description: Post created
        '401':
          description: Unauthorized
        '404':
          description: Thread not found
  /threads/{id}/approve:
    post:
      summary: Approve thread (mod/admin)
      tags: [Threads]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved
        '400':
          description: Invalid transition
        '403':
          description: Forbidden
  /threads/{id}/hide:
    post:
      summary: Hide or unhide thread (mod/admin)
      tags: [Threads]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                hide:
                  type: boolean
                  default: true
                reason:
                  type: string
      responses:
        '200':
          description: Hidden/unhidden
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /users/{id}:
    get:
      summary: Get user profile
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile (passwordHash excluded)
        '404':
          description: Not found
    patch:
      summary: Update own profile
      tags: [Users]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                theme:
                  type: string
      responses:
        '200':
          description: Updated user
        '403':
          description: Forbidden
        '404':
          description: Not found
  /auth/register:
    post:
      summary: Register
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  pattern: '^[A-Za-z0-9]{3,24}$'
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                honeypot:
                  type: string
                  description: Must be empty (bot trap)
      responses:
        '201':
          description: Registered successfully
        '400':
          description: Validation error
        '409':
          description: Email or username already in use
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [usernameOrEmail, password]
              properties:
                usernameOrEmail:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Logged in
        '401':
          description: Invalid credentials
        '403':
          description: Account banned
  /auth/logout:
    post:
      summary: Logout
      tags: [Auth]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Logged out
        '401':
          description: Unauthorized
...
  /admin/users/{id}:
    patch:
      summary: Update user role/trust (admin only)
      tags: [Admin]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [user, moderator, admin]
                trust:
                  type: string
                  enum: [new, member, trusted, banned]
                isDeleted:
                  type: boolean
                reason:
                  type: string
      responses:
        '200':
          description: Updated user
        '403':
          description: Forbidden
        '404':
          description: Not found
...